add separate job listings servlet
