+++ walk-in-interview/src/main/java/com/google/job/servlets/JobServlet.java
@@ -32,11 +32,6 @@ public final class JobServlet extends HttpServlet {
     private static final String JOB_ID_FIELD = "jobId";
     private static final long TIMEOUT = 5;
 
-    public static final String SORT_BY_PARAM = "sortBy";
-    public static final String ORDER_PARAM = "order";
-    public static final String PAGE_SIZE_PARAM = "pageSize";
-    public static final String PAGE_INDEX_PARAM = "pageIndex";
-
     private JobsDatabase jobsDatabase;
 
     @Override
@@ -56,20 +51,7 @@ public final class JobServlet extends HttpServlet {
 
     @Override
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        try {
-            Filter sortBy = parseSortBy(request);
-            Order order = parseOrder(request);
-            int pageSize = parsePageSize(request);
-            int pageIndex = parsePageIndex(request);
-
-            JobPage jobPage = fetchJobPageDetails(sortBy, order, pageSize, pageIndex);
-
-            String json = new Gson().toJson(jobPage);
-            response.setContentType("application/json;");
-            response.getWriter().println(json);
-        } catch(IllegalArgumentException | ServletException | ExecutionException | TimeoutException e) {
-            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-        }
+        // TODO(issue/60): get individual job post from jobId
     }
 
     @Override
@@ -113,103 +95,6 @@ public final class JobServlet extends HttpServlet {
             System.err.println("Error occur: " + e.getCause());
             // Sends the fail status code in the response
             response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-        }
-    }
-
-    /**
-     * Returns the JobPage object.
-     *
-     * @param sortBy The sorting of the list of jobs.
-     * @param order The ordering of the sorting.
-     * @param pageSize The number of job listings to be returned.
-     * @param pageIndex The page which we are on (pagination).
-     * @return JobPage object with all the details for the GET response.
-     */
-    private JobPage fetchJobPageDetails(Filter sortBy, Order order, int pageSize, int pageIndex)
-            throws ServletException, ExecutionException, TimeoutException {
-        try {
-            return this.jobsDatabase.fetchJobPage(SingaporeRegion.ENTIRE, sortBy, order, pageSize, pageIndex)
-                    .get(TIMEOUT, TimeUnit.SECONDS);
-        } catch (InterruptedException | IOException e) {
-            throw new ServletException(e);
-        }
-    }
-
-    /**
-     * Returns the sorting as a Filter enum.
-     *
-     * @param request From the GET request.
-     * @return Filter enum given the id in the request params.
-     * @throws IllegalArgumentException if the id is invalid.
-     */
-    public static Filter parseSortBy(HttpServletRequest request) throws IllegalArgumentException {
-        String sortById = (String) request.getParameter(SORT_BY_PARAM);
-
-        if (sortById == null || sortById.isEmpty()) {
-            throw new IllegalArgumentException("sort by param should not be null or empty");
-        }
-
-        return Filter.getFromId(sortById); // IAE may be thrown
-    }
-
-    /**
-     * Returns the ordering as an Order enum.
-     *
-     * @param request From the GET request.
-     * @return Order enum given the id in the request params.
-     * @throws IllegalArgumentException if the id is invalid.
-     */
-    public static Order parseOrder(HttpServletRequest request) throws IllegalArgumentException {
-        String orderId = (String) request.getParameter(ORDER_PARAM);
-
-        if (orderId == null || orderId.isEmpty()) {
-            throw new IllegalArgumentException("order param should not be null or empty");
-        }
-
-        return Order.getFromId(orderId); // IAE may be thrown
-    }
-
-    /**
-     * Returns the page size as an int.
-     *
-     * @param request From the GET request.
-     * @return the page size.
-     * @throws IllegalArgumentException if the page size is invalid.
-     */
-    public static int parsePageSize(HttpServletRequest request) throws IllegalArgumentException {
-        String pageSizeStr = (String) request.getParameter(PAGE_SIZE_PARAM);
-
-        if (pageSizeStr == null || pageSizeStr.isEmpty()) {
-            throw new IllegalArgumentException("page size param should not be null or empty");
-        }
-
-        try {
-            int pageSize =  Integer.parseInt(pageSizeStr);
-            return pageSize;
-        } catch(NumberFormatException e) {
-            throw new IllegalArgumentException("page size param should be an int");
-        }
-    }
-
-    /**
-     * Returns the page index as an int.
-     *
-     * @param request From the GET request.
-     * @return the page index.
-     * @throws IllegalArgumentException if the page index is invalid.
-     */
-    public static int parsePageIndex(HttpServletRequest request) throws IllegalArgumentException {
-        String pageIndexStr = (String) request.getParameter(PAGE_INDEX_PARAM);
-
-        if (pageIndexStr == null || pageIndexStr.isEmpty()) {
-            throw new IllegalArgumentException("page index param should not be null or empty");
-        }
-
-        try {
-            int pageIndex =  Integer.parseInt(pageIndexStr);
-            return pageIndex;
-        } catch(NumberFormatException e) {
-            throw new IllegalArgumentException("page index param should be an int");
         }
     }
